Changed folder structure and renamed according to main project.
