+++ lang/csharp/Eclipse.eCAL.Core/include/clr_common.h
@@ -5,9 +5,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,19 +17,21 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include "ecal_clr_monitoring.h"
-#include "ecal_clr_common.h"
+#pragma once
 
-#include <ecal/core.h>
-#include <ecal/log.h>
-#include <ecal/monitoring.h>
+/**
+ * @file  clr_common.h
+**/
 
-using namespace Eclipse::eCAL::Core;
-using namespace Internal;
+#include <string>
 
-array<Byte>^ Monitoring::GetMonitoring(MonitoringEntity entities)
+using namespace System;
+
+namespace Internal
 {
-  std::string monitoring;
-  ::eCAL::Monitoring::GetMonitoring(monitoring, static_cast<unsigned int>(entities));
-  return StlStringToByteArray(monitoring);
+  System::String^ StlStringToString(const std::string& ss_);
+  std::string     StringToStlString(System::String^ s_);
+
+  std::string  ByteArrayToStlString(array<Byte>^ a_);
+  array<Byte>^ StlStringToByteArray(const std::string& ss_);
 }
