+++ lang/csharp/Eclipse.eCAL.Core/include/clr_log.h
@@ -20,7 +20,7 @@
 #pragma once
 
 /**
- * @file  ecal_clr_monitoring.h
+ * @file  clr_log.h
 **/
 
 using namespace System;
@@ -29,34 +29,38 @@ namespace Eclipse {
   namespace eCAL {
     namespace Core {
 
-      // Managed representation of the native monitoring entities bitmask.
-      [Flags]
-      public enum class MonitoringEntity : unsigned int
+      // Expose a managed log level enum with values matching the native API.
+      [System::Flags]
+      public enum class LogLevel : unsigned char
       {
-        None       = 0x000,
-        Publisher  = 0x001,
-        Subscriber = 0x002,
-        Server     = 0x004,
-        Client     = 0x008,
-        Process    = 0x010,
-        Host       = 0x020,
-        All        = Publisher | Subscriber | Server | Client | Process | Host
+        None    = 0,
+        Info    = 1,    // native: log_level_info
+        Warning = 2,    // native: log_level_warning
+        Error   = 4,    // native: log_level_error
+        Fatal   = 8,    // native: log_level_fatal
+        All     = 255   // native: log_level_all
       };
 
       /**
-       * @brief eCAL monitoring class.
-       **/
-      public ref class Monitoring
+       * @brief eCAL logging class.
+      **/
+      public ref class Logging
       {
       public:
         /**
-         * @brief Get host, process and topic monitoring as raw message bytes.
+         * @brief Log a message with a specified log level.
          *
-         * @param entities Specifies which entities to include (default is All).
-         * 
-         * @return The monitoring message as a byte array.
+         * @param level    The log level.
+         * @param message  The message string to log.
+        **/
+        static void Log(LogLevel level, System::String^ message);
+
+        /**
+         * @brief Get global log message as raw message bytes.
+         *
+         * @return The logging message as a byte array.
          **/
-        static array<Byte>^ GetMonitoring(MonitoringEntity entities);
+        static array<Byte>^ GetLogging();
       };
 
     } // namespace Core
