+++ lang/csharp/Eclipse.eCAL.Core/src/clr_client.cpp
@@ -19,10 +19,10 @@
 
 #include <msclr/marshal_cppstd.h>
 
-#include "ecal_clr_client.h"
+#include "service/clr_client.h"
+#include "service/clr_types.h"
 
-#include "ecal_clr_common.h"
-#include "ecal_clr_types_service.h"
+#include "clr_common.h"
 
 #include <vector>
