+++ lang/csharp/Eclipse.eCAL.Core/src/clr_monitoring.cpp
@@ -17,27 +17,19 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include "ecal_clr_log.h"
-#include "ecal_clr_common.h"
+#include "clr_monitoring.h"
+#include "clr_common.h"
 
+#include <ecal/core.h>
 #include <ecal/log.h>
+#include <ecal/monitoring.h>
 
 using namespace Eclipse::eCAL::Core;
 using namespace Internal;
 
-void Logging::Log(LogLevel level, System::String^ message)
+array<Byte>^ Monitoring::GetMonitoring(MonitoringEntity entities)
 {
-  // Convert the managed string to a native std::string
-  std::string nativeMsg = StringToStlString(message);
-  
-  // Cast the managed LogLevel to the native eCAL::Logging::eLogLevel.
-  // The values match, so a static_cast is sufficient.
-  ::eCAL::Logging::Log(static_cast<::eCAL::Logging::eLogLevel>(level), nativeMsg);
-}
-
-array<Byte>^ Logging::GetLogging()
-{
-  std::string logging;
-  ::eCAL::Logging::GetLogging(logging);
-  return StlStringToByteArray(logging);
+  std::string monitoring;
+  ::eCAL::Monitoring::GetMonitoring(monitoring, static_cast<unsigned int>(entities));
+  return StlStringToByteArray(monitoring);
 }
