+++ lang/csharp/Eclipse.eCAL.Core/src/clr_server.cpp
@@ -17,10 +17,10 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include "ecal_clr_server.h"
+#include "service/clr_server.h"
 
-#include "ecal_clr_common.h"
-#include "ecal_clr_types.h"
+#include "clr_common.h"
+#include "clr_types.h"
 
 #include <sstream>
