+++ lang/csharp/Eclipse.eCAL.Core/src/clr_time.cpp
@@ -17,8 +17,8 @@
  * ========================= eCAL LICENSE =================================
 */
 
-#include "ecal_clr_time.h"
-#include "ecal_clr_common.h"
+#include "clr_time.h"
+#include "clr_common.h"
 
 #include <ecal/time.h>
